# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Spec Doc API
  version: "1.0.0"
  description: >
    Phase 1 contract — health check, projects CRUD, and context
    read/write for four named keys (builder, principles, codebase,
    references). This file is the build artifact; flask/dtos/ and
    mock_server.py are derived from it.
servers:
  - url: http://localhost:3101
    description: Flask development server
  - url: http://localhost:3102
    description: Mock server (Task 3)
paths:
...
  /api/projects:
    get:
      summary: List all projects, newest first
      operationId: listProjects
      responses:
        "200":
          description: Array of project summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectSummary"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      summary: Create a new project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectCreateRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectCreateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/projects/{id}:
    get:
      summary: Get project detail with full file contents
      operationId: getProject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      summary: Delete a project
      operationId: deleteProject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/projects/{id}/coherence:
    post:
      summary: Run the multi-doc coherence pass against a project directory
      operationId: runProjectCoherence
      tags: [quality]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Coherence flags and summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoherenceResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/projects/{id}/files/{filename}:
    put:
      summary: Update a file within a project
      operationId: updateProjectFile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: filename
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileUpdateRequest"
      responses:
        "200":
          description: File updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/projects/{project_id}/files/{filename}/history:
    get:
      summary: Commit history for a project file
      operationId: getFileHistory
      tags: [projects]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
        - name: filename
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileHistoryResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/projects/{project_id}/files/{filename}/diff:
    get:
      summary: Unified diff between two SHAs for a project file
      operationId: getFileDiff
      tags: [projects]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
        - name: filename
          in: path
          required: true
          schema:
            type: string
        - name: from_sha
          in: query
          required: true
          schema:
            type: string
        - name: to_sha
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileDiffResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/projects/{project_id}/files/{filename}/revert:
    post:
      summary: Revert a project file to a specific commit SHA (forward commit)
      operationId: revertFile
      tags: [projects]
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
        - name: filename
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileRevertRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileRevertResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/projects/{id}/repair:
    post:
      summary: Write any missing deterministic structural files for a project
      operationId: repairProject
      tags: [projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: >
            Repair complete. repaired lists filenames written;
            empty array means no files were missing.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepairProjectResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/auth/login:
    post:
      summary: Request a magic-link email for the supplied address.
      operationId: requestMagicLink
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MagicLinkRequest'
      responses:
        '202':
          description: Magic-link dispatch accepted by Neon Auth.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MagicLinkResponse'
  /api/auth/verify:
    post:
      summary: Exchange a one-time magic-link token for a session JWT.
      operationId: verifyMagicLink
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyRequest'
      responses:
        '200':
          description: Token exchanged for JWT plus hydrated user row.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyResponse'
...
  /api/auth/me:
    get:
      summary: Return the currently authenticated user row.
      operationId: getCurrentUser
      tags: [auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current authenticated user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
components:
  schemas:
...
    SpecSummary:
      type: object
      required: [filename, label]
      properties:
        filename:
          type: string
        label:
          type: string
    SpecDetail:
      type: object
      required: [filename, label, content]
      properties:
        filename:
          type: string
        label:
          type: string
        content:
          type: string
    ProjectSummary:
      type: object
      required: [id, name, createdAt, specs]
      properties:
        id:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        specs:
          type: array
          items:
            $ref: "#/components/schemas/SpecSummary"
    ProjectDetail:
      type: object
      required: [id, name, createdAt, specs]
      properties:
        id:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        specs:
          type: array
          items:
            $ref: "#/components/schemas/SpecDetail"
...
    CoherenceFlag:
      type: object
      required: [rule, severity, message]
      properties:
        rule:
          type: string
          description: Invariant identifier (e.g. symbol_uniqueness, content_routing).
        severity:
          type: string
          enum: [error, warning]
        message:
          type: string
        line:
          type: integer
          description: 1-based line number in the source document, when applicable.
    CoherenceResponse:
      type: object
      required: [flags, summary]
      properties:
        flags:
          type: array
          items:
            $ref: "#/components/schemas/CoherenceFlag"
        summary:
          type: string
          description: 'Human-readable count string, e.g. "3 flags: 1 error, 2 warnings."'
    ProjectCreateRequest:
      type: object
      required: [name, files]
      properties:
        name:
          type: string
        files:
          type: array
          items:
            $ref: "#/components/schemas/FileInput"
    FileInput:
      type: object
      required: [filename, content]
      properties:
        filename:
          type: string
        content:
          type: string
    ProjectCreateResponse:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
    FileUpdateRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
    CommitEntry:
      type: object
      required: [sha, message, author, timestamp]
      properties:
        sha:
          type: string
        message:
          type: string
        author:
          type: string
        timestamp:
          type: string
          format: date-time
    FileHistoryResponse:
      type: object
      required: [history]
      properties:
        history:
          type: array
          items:
            $ref: "#/components/schemas/CommitEntry"
    FileDiffResponse:
      type: object
      required: [diff]
      properties:
        diff:
          type: string
    FileRevertRequest:
      type: object
      required: [to_sha]
      properties:
        to_sha:
          type: string
        path:
          type: string
          description: Optional alias for filename; the route's path parameter is canonical.
    FileRevertResponse:
      type: object
      required: [sha, message]
      properties:
        sha:
          type: string
        message:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    RepairProjectResponse:
      type: object
      required: [repaired]
      properties:
        repaired:
          type: array
          items:
            type: string
          description: >
            Filenames written during this repair run.
            Empty array means all structural files were already present.
...
    MagicLinkRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          description: Address that will receive the magic-link email.
    MagicLinkResponse:
      type: object
      required: [request_id]
      properties:
        request_id:
          type: string
          description: >
            Opaque identifier returned by Neon Auth for the dispatched
            magic-link request. Useful for support diagnostics.
    VerifyRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: One-time token extracted from the magic-link URL fragment.
    VerifyResponse:
      type: object
      required: [jwt, user]
      properties:
        jwt:
          type: string
          description: Session JWT issued by Neon Auth. Pass as Authorization Bearer.
        user:
          $ref: '#/components/schemas/MeResponse'
    MeResponse:
      type: object
      required: [id, email, auth_user_id]
      properties:
        id:
          type: integer
          description: Spec-doc internal user row id.
        email:
          type: string
          description: User email (mirrors the Neon Auth claim).
        auth_user_id:
          type: string
          description: >
            Stable identity from Neon Auth (sub claim). FK anchor for every
            tenant-scoped table in saas-persistence.
  responses:
    BadRequest:
      description: Missing or invalid request fields
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Neon Auth JWT issued at login. Pass as Authorization: Bearer <token>.
        Verified by the require_auth decorator (implemented in Task 2).